Remove fp32 deps from irr

+++ ptlflow/models/irr/pwcnet_irr.py
@@ -103,9 +103,7 @@ class IRRPWCNetIRR(BaseModel):
         ) = x1_pyramid[0].size()
         init_dtype = x1_pyramid[0].dtype
         init_device = x1_pyramid[0].device
-        flow = torch.zeros(
-            b_size, 2, h_x1, w_x1, dtype=init_dtype, device=init_device
-        ).float()
+        flow = torch.zeros(b_size, 2, h_x1, w_x1, dtype=init_dtype, device=init_device)
 
         for l, (x1, x2) in enumerate(zip(x1_pyramid, x2_pyramid)):
             # warping
